requirements.txt
langdetect==1.0.9
pdf2image==1.17.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-docx==1.2.0
scikit-learn==1.7.0
spacy==3.8.7
streamlit==1.46.0


